+++ src/gitutils/_git-release-prod.sh
@@ -45,12 +45,16 @@ GIT_EDITOR=:
 GBV=$(bump-changelog -b -m)
 if [ "$?" -eq 0 ] && [ -n "$GBV" ]; then
     zz_log s "Version & CHANGELOG updated to: {B $GBV}"
-    git commit -am "chore(release): $GBV"
+    
+    if ! git commit -am "chore(release): $GBV"; then
+        zz_log e "Cannot commit version & CHANGELOG"
+        exit 1
+    fi
+
     if git flow $flow finish $name --push; then
         zz_log s "Release finished: {B $GBV}"
         rm -f .git/RELEASE
     else
-        git undo
         zz_log e "Cannot finish release. CHANGELOG & VERSION are not updated."
     fi
 else